Levrage BigInt to represent int64/uint64 (#1030)
Due to [double precision floating point format](https://en.wikipedia.org/wiki/Double_precision_floating-point_format),  JavaScript can only safely represent integers [between](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/MAX_SAFE_INTEGER) -(2^53 – 1) to 2^53 – 1, meanwhile, `Number.MAX_SAFE_INTEGER` is defined to represent the maximum safe integer in JavaScript, which is 9007199254740991. Per ROS2 message types, it has [int64](https://github.com/ros2/common_interfaces/blob/rolling/std_msgs/msg/Int64.msg)/[uint64](https://github.com/ros2/common_interfaces/blob/rolling/std_msgs/msg/UInt64.msg), which may be out of the range above.

For current implementation, we leverage [ref](https://github.com/node-ffi-napi/ref-napi) to read/write values from `int64_t` and `uint64_t` data of C++, which can be type of:
1. `number`: when value in [`-Number.MAX_SAFE_INTEGER`  `Number.MAX_SAFE_INTEGER`].
2. `string`: otherwise.

This brings the problem that the type of int64/uint64 be ununified in JavaScript. Along with [BigInt](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/BigInt) introduced, it becomes possible to make the type consistent.

This patch implements to leverage `BigInt` to represent int64/uint64, including:
1. Update `message.dot` to use `BigInt` when hitting int64/uint64.
2. Update `rosidl_parser.js` to convert  int64/uint64 to string in JSON object.
3. Update `message_translator.js` to add `bigint` as primitive.
4. Update `parameter.js` to use `BigInt` for integer and its `ts` definition `parameter.d.ts`.
5. Update tests targeting on messages that include int64/uint64.

Fix: #836, #1025

+++ test/test-service-introspection.js
@@ -86,8 +86,8 @@ describe('service introspection', function () {
     }
 
     this.request = {
-      a: Math.floor(Math.random() * 100),
-      b: Math.floor(Math.random() * 100),
+      a: BigInt(Math.floor(Math.random() * 100)),
+      b: BigInt(Math.floor(Math.random() * 100)),
     };
 
     this.eventQueue = [];
